COMP: Fixed warnings.

+++ Servers/ServerManager/vtkSMSILModel.cxx
@@ -450,7 +450,8 @@ void vtkSMSILModel::UpdatePropertyValue(vtkSMStringVectorProperty* svp)
     values[2*cc] = this->GetName(*iter);
     values[2*cc+1] = check_states[this->GetCheckStatus(*iter)];
     }
-  svp->SetElements(leaf_ids.size()*2, values);
+  svp->SetElements(
+    static_cast<unsigned int>(leaf_ids.size())*2, values);
   delete[] values;
   this->BlockUpdate = false;
 }
